CI 补上 Windows 目标的 clippy，并清掉它暴露的 4 处空转换
Windows 专有代码此前从没被 lint 过：lint job 只跑本机（macOS）目标，被
#[cfg(windows)] 挡住的文件它一行都看不到；cross-check 虽然编译 Windows 目标，
用的却是 cargo check，抓不到 clippy 级别的问题。于是 sockopt_win.rs、
dialog_win_task.rs、托盘的 Windows 分支全在盲区里。

发现方式本身就说明问题：上一轮为了验证新写的 sockopt_win.rs 才手动跑了一次
带 --target 的 clippy，顺带发现 dialog_win_task.rs 里攒了 4 处
unnecessary_cast——windows-sys 0.60 的 IDOK/IDNO/IDYES/IDCANCEL 已经是 i32，
再 `as i32` 是空转换。没人跑就没人知道，这种事不该靠碰巧。

cross-check 的 Windows 那步从 cargo check 换成 cargo clippy。clippy 本身就做
完整类型检查，原来那步要拦的"另一个平台编译不过"照样拦得住，多出来的只有
lint，而这个 job 已经装好目标、有独立缓存，比在 lint job 里另配一份便宜。

Intel macOS 那步保持 cargo check：与本机同属 macOS，cfg 分支一致，lint 结果
不会有差别，跑 clippy 是白花时间。

注意 dialog_win_task.rs 第 145 行的 `i as i32` 是真转换（usize → i32），留着。

+++ crates/clipsync-app/src/dialog_win_task.rs
@@ -108,7 +108,7 @@ pub fn ask_action(title: &str, body: &str, action: &str) -> Result<bool> {
     cfg.pButtons = buttons.as_ptr();
     // TDCBF_OK_BUTTON = 1
     cfg.dwCommonButtons = 1;
-    cfg.nDefaultButton = IDOK as i32;
+    cfg.nDefaultButton = IDOK;
 
     Ok(run(&cfg)? == FIRST_BUTTON_ID)
 }
@@ -124,8 +124,8 @@ pub fn confirm(title: &str, body: &str) -> Result<bool> {
     cfg.pszContent = pcwstr(&body_w);
     // TDCBF_YES_BUTTON | TDCBF_NO_BUTTON = 2 | 4
     cfg.dwCommonButtons = 2 | 4;
-    cfg.nDefaultButton = IDNO as i32;
-    Ok(run(&cfg)? == IDYES as i32)
+    cfg.nDefaultButton = IDNO;
+    Ok(run(&cfg)? == IDYES)
 }
 
 /// 列表选择，用 Command Links 呈现——每个选项是一个大按钮，比列表框好按也好看。
@@ -157,7 +157,7 @@ pub fn choose(title: &str, body: &str, items: &[String]) -> Result<Option<usize>
     cfg.dwCommonButtons = 8;
 
     let pressed = run(&cfg)?;
-    if pressed == IDCANCEL as i32 {
+    if pressed == IDCANCEL {
         return Ok(None);
     }
     let idx = pressed - FIRST_BUTTON_ID;
